CAS-MM Sync: Now comparing user data in memory rather than using user_load()
This makes the assumption that the user casmmsync_attributes stored in the
user.data field (on the last sync/login) are an accurate representation of the
state of the user. It would be possible for this not to be the case if the user's
profile fields were edited through the Drupal UI. That said, the next time that
the user logs in the full user object and all of its profile fields will be directly
compared and reset to their source values, so this isn't too much of a concern.

With this change a single query suffices to load all of the data for all users
that have not had their attributes changed. All comparisons are done in memory
and are very fast.