Midd CAS: Filtering out users on login or sync who aren't members of one of our allowed groups.
For now this is hard-coded to the "institution" group, but could be made an admin option
if we want it to be different for different sites.

With this change, users who aren't in the "institution" group will not be added to
the Drupal user database.

This maps to D6 commits:
a99378a4e78a4e8c0f3dd0eb3b11cf6f903d9d9b
8bc8bbfdb5741ef23c5f3fba4fa2bac735699281